\name{print.knn}
\alias{print.knn}
\title{Print a knn object}
\description{Print a knn objectdata}
\details{This prints a knn object which will normally have been
created by \R{knn.var}}
\usage{print.knn (x, ...)}
\arguments{
\item{x} {knn.var model created by \R{knn.var}}
\item{...} {Other arguments, currently ignored}}

\keywords{print}
\references{Buttrey and Karo, 2002}
\author{Sam Buttrey \email{buttrey@nps.navy.mil}}
\seealso{\code{link{knn.var}}, \code{\link{knnTree}}}
