\name{print.knnTree}
\alias{print.knnTree}
\title{Print a knnTree object}
\description{Print a knnTree object}
\details{This prints a knnTree object which will normally have been
created by \R{knnTree}}
\usage{print.knnTree (x, ...)}
\arguments{
\item{x} {knnTree model created by \R{knnTree}}
\item{...} {Other arguments, currently ignored}}


\keywords{print}
\references{Buttrey and Karo, 2002}
\author{Sam Buttrey \email{buttrey@nps.navy.mil}}
\seealso{\code{link{knn.var}}, \code{\link{knnTree}}}
